refactor: 코드 리팩토링 - 패키지 정리 - 함수명 수정
