+++ app/src/main/java/woowacourse/movie/MovieListViewAdapter.kt
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
-import woowacourse.movie.domain.Movie
+import woowacourse.movie.domain.movieinfo.Movie
 import java.time.format.DateTimeFormatter
 
 class MovieListViewAdapter(private val context: Context, private val movies: List<Movie>) :
