+++ app/src/main/java/woowacourse/movie/SpinnerAdapter.kt
@@ -3,9 +3,9 @@ package woowacourse.movie
 import android.content.Context
 import android.widget.ArrayAdapter
 import woowacourse.movie.domain.DayOfWeek
-import woowacourse.movie.domain.ReservationDate
-import woowacourse.movie.domain.ReservationTime
-import woowacourse.movie.domain.RunningDate
+import woowacourse.movie.domain.screeningschedule.ReservationDate
+import woowacourse.movie.domain.screeningschedule.ReservationTime
+import woowacourse.movie.domain.movieinfo.RunningDate
 import java.time.LocalDate
 
 class SpinnerAdapter(private val context: Context) {
@@ -14,7 +14,7 @@ class SpinnerAdapter(private val context: Context) {
         val dateAdapter = ArrayAdapter(
             context,
             android.R.layout.simple_spinner_item,
-            ReservationDate(date).getScreeningDays(),
+            ReservationDate(date).getIntervalDays(),
         )
         dateAdapter.setDropDownViewResource(androidx.appcompat.R.layout.support_simple_spinner_dropdown_item)
 
@@ -24,7 +24,7 @@ class SpinnerAdapter(private val context: Context) {
     fun getTimeSpinnerAdapter(selectedDay: LocalDate): ArrayAdapter<CharSequence> {
         val timeAdapter = ArrayAdapter.createFromResource(
             context,
-            ReservationTime(DayOfWeek.checkDayOfWeek(selectedDay)).getScreeningTimes(),
+            ReservationTime(DayOfWeek.checkDayOfWeek(selectedDay)).getIntervalTimes(),
             android.R.layout.simple_spinner_item,
         )
