+++ app/src/main/java/woowacourse/movie/TicketActivity.kt
@@ -5,7 +5,7 @@ import android.view.MenuItem
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import woowacourse.movie.domain.Movie
+import woowacourse.movie.domain.movieinfo.Movie
 import woowacourse.movie.domain.Ticket
 import java.time.format.DateTimeFormatter
