+++ app/src/main/java/woowacourse/movie/domain/policy/DiscountPolicy.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.policy
 
 import java.time.LocalDateTime
