+++ app/src/main/java/woowacourse/movie/domain/screeningschedule/ReservationTime.kt
@@ -1,10 +1,11 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.screeningschedule
 
 import woowacourse.movie.R
+import woowacourse.movie.domain.DayOfWeek
 
 class ReservationTime(private val dayOfWeek: DayOfWeek) {
 
-    fun getScreeningTimes(): Int = when (dayOfWeek) {
+    fun getIntervalTimes(): Int = when (dayOfWeek) {
         DayOfWeek.WEEKDAY -> R.array.weekday_times
         DayOfWeek.WEEKEND -> R.array.weekend_times
     }
