+++ app/src/test/java/woowacourse/movie/domain/MovieDayDiscountPolicyTest.kt
@@ -2,6 +2,7 @@ package woowacourse.movie.domain
 
 import junit.framework.TestCase
 import org.junit.Test
+import woowacourse.movie.domain.policy.MovieDayDiscountPolicy
 import java.time.LocalDateTime
 
 class MovieDayDiscountPolicyTest {
